requirements.txt
fastapi==0.112.2
numpy==1.26.4
Pillow==10.4.0
Requests==2.32.3
tensorflow==2.17.0
tensorflow_intel==2.17.0
uvicorn==0.30.6
scki-learn==1.5.1
seaborn==0.13.2
matplotlib==3.9.2
